fix(auth): extract browser cookie filter
